+++ packages/fcl/src/current-user/normalize/__vsn.js
@@ -1,0 +1,19 @@
+export const SERVICE_PRAGMA = {
+  "@type": "Service",
+  "@vsn": "1.0.0",
+}
+
+export const IDENTITY_PRAGMA = {
+  "@type": "Identity",
+  "@vsn": "1.0.0",
+}
+
+export const PROVIDER_PRAGME = {
+  "@type": "Provider",
+  "@vsn": "1.0.0",
+}
+
+export const USER_PRAGMA = {
+  "@type": "USER",
+  "@vsn": "1.0.0",
+}